requirements.txt
torch>=2.1.2
scipy>=1.15.3
Ninja==1.10.2
gradio>=3.35.2
imageio-ffmpeg>=0.6.0
huggingface_hub>=1.3.2
hf_transfer>=0.1.9
pyopengl>=3.1.10
imgui==2.0.0
glfw==2.6.1
pillow>=12.1.0
torchvision>=0.16.2
imageio>=2.37.2
numpy>=1.26.4
matplotlib>=3.10.8
tqdm>=4.67.1
pyspng-seunglab>=1.1.2
click>=8.3.1
